refactor(scroll): move scrollToContent method inside a utils file
