+++ src/app/features/components/back-to-top/back-to-top.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { scrollToContent } from '../../../shared/components/menus/menu.conf';
+import { scrollToContent } from '../../../shared/utils/scroll';
 
 @Component({
   selector: 'app-back-to-top',
