+++ src/app/shared/components/menus/menu.conf.ts
@@ -26,14 +26,3 @@ export const links: LinkType[] = [
     isMounted: false,
   },
 ];
-
-export const scrollToContent = (id: string): void => {
-  const targetId = id.split('#')[1];
-  const element = document.getElementById(targetId);
-
-  element?.scrollIntoView({
-    behavior: 'smooth',
-    block: 'center',
-    inline: 'nearest',
-  });
-};
